Search groups through pdf

+++ Backend/ProjekatSVT/src/main/java/com/example/ProjekatSVT/searchmodel/GroupIndex.java
@@ -10,6 +10,7 @@ import org.springframework.data.elasticsearch.annotations.FieldType;
 import org.springframework.data.elasticsearch.annotations.Setting;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
@@ -19,7 +20,7 @@ import java.time.LocalDate;
 public class    GroupIndex {
 
     @Id
-    private Long id;
+    private Integer id;
 
     @Field(type = FieldType.Text, store = true, name = "name", analyzer = "serbian_simple", searchAnalyzer = "serbian_simple")
     private String name;
